Fix #1807914: Title showing as "Unknown" in device list for Kobo devices
Fix for https://bugs.launchpad.net/calibre/+bug/1807914
With the last set of changes, if the metadata management was not set to
automatic, books sent to the device would have the title showing as
"Unknown" in the device list.

Also bumping supported firmware version to 4.12.12111 which has just
been released.