PDF Output: Fix incorrect rendering when the same font is used at multiple sizes in the document
Chromium produces width arrays in the font descriptors that contain
different values for the same glyph in different font subsets, so rather
than merging the arrays, use the actual width values from the font.

As per the PDF spec the values in these arrays must match the underlying
font anyway.