build: detect mathjax directory with or without es5/ subdir
The release tarballs contain an es5/ directory with the actual mathjax
code. However, when building against a system mathjax it is likely the
es5/ directory is skipped and the directory name is just
/usr/share/mathjax/

Detect this by moving the es5/ component to the definition of the
mathjax source root, but only appending it if the --path-to-mathjax
actually contains such a directory.