Fix for Bug #1909738: Strangeness with hierarchical searches
Because it is useful I decided to continue to permit hierarchy for saved searches even though behavior will be a bit strange.

The following behaviors exist and will not be changed.
- Renames will not rename all items in the the hierarchy.
- Intermediate nodes that aren't 'real' searches will not offer the rename option.
- Drag & drop is disabled.

In addition I changed the behavior of "Add search" in "Manage saved searches". It used to act like a rename if the target name already exists, throwing away the old search expression. It now refuses.